+++ tests/heading_code_spec.lua
@@ -6,26 +6,26 @@ describe('heading_code.md', function()
     it('default', function()
         util.setup('demo/heading_code.md')
 
-        local expected = {}
+        local expected, row = {}, util.row()
 
         vim.list_extend(expected, {
-            util.heading(0, 1),
-            util.heading(2, 3),
-            util.heading(4, 4),
-            util.heading(6, 5),
-            util.heading(8, 6),
+            util.heading(row:get(), 1),
+            util.heading(row:increment(2), 3),
+            util.heading(row:increment(2), 4),
+            util.heading(row:increment(2), 5),
+            util.heading(row:increment(2), 6),
         })
 
-        table.insert(expected, util.link(10, 0, 21, 'image'))
+        table.insert(expected, util.link(row:increment(2), 0, 21, 'image'))
 
         vim.list_extend(expected, {
-            util.code_row(12, 0),
-            util.code_language(12, 3, 9, 'python'),
+            util.code_row(row:increment(2), 0),
+            util.code_language(row:get(), 3, 9, 'python'),
         })
-        for i = 13, 21 do
-            table.insert(expected, util.code_row(i, 0))
+        for _ = 13, 21 do
+            table.insert(expected, util.code_row(row:increment(), 0))
         end
-        table.insert(expected, util.code_below(22, 0))
+        table.insert(expected, util.code_below(row:increment(), 0))
 
         local actual = util.get_actual_marks()
         util.marks_are_equal(expected, actual)